# docs/swagger-docs/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/activity": {
      "get": {
        "tags": [
          "UserActivity-Controller"
        ],
        "operationId": "getAllUserActivitiesBasedOnEventType",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          },
          {
            "name": "arg0",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserActivityRequest"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActivityRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "activity",
                  "attributes": {
                    "eventType": "sign-in"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOListUserActivity"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UserActivity-Controller"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "operationId": "createUserActivity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActivityRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "activity",
                  "attributes": {
                    "eventType": "Sign-in",
                    "userId": "10",
                    "ipAddress": "192.168.9.5",
                    "eventDetails": "User signed in"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "User-Controller"
        ],
        "operationId": "retrieveUser",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User-Controller"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {}
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/activity/{param}": {
      "get": {
        "tags": [
          "UserActivity-Controller"
        ],
        "operationId": "getUserActivityForSpecificUser",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          },
          {
            "name": "param",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arg1",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestDTO"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActivityRequest"
              },
              "example": {
                "data": {
                  "id": "13453",
                  "type": "activity",
                  "attributes": {
                    "startDate": "2024-10-10",
                    "endDate": "2024-10-18",
                    "eventType": "Sign-in",
                    "email": "vishwasransingh@gmail.com",
                    "mobile": "7743898263"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/partners/banks/signIn": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "signIn",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "email": "partner@example.com",
                    "mobile": "9876543210",
                    "password": "password123",
                    "emailOtp": "123456",
                    "mobileOtp": "654321",
                    "ipAddress": "192.168.1.1"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Partner signed in successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/password/reset": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "resetPassword",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "email": "partner@example.com",
                    "mobile": "9876543210",
                    "emailOtp": "123456",
                    "mobileOtp": "654321",
                    "ipAddress": "192.168.1.1"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/profile": {
      "put": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "updateProfile",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "email": "partner@example.com",
                    "mobile": "9876543210",
                    "password": "newPassword123",
                    "emailOtp": "123456",
                    "mobileOtp": "654321",
                    "ipAddress": "192.168.1.1",
                    "companyName": "Partner Corp",
                    "companyAddress": "123 Partner Street",
                    "dateOfIncorporation": "2000-01-01",
                    "license": "ABC123",
                    "cin": "XYZ789",
                    "gstNo": "12AABCU1234K1ZV"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Partner profile updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/signUp": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "signUp",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "email": "partner@example.com",
                    "mobile": "9876543210",
                    "password": "password123",
                    "emailOtp": "123456",
                    "mobileOtp": "654321",
                    "ipAddress": "192.168.1.1",
                    "companyName": "Partner Corp",
                    "companyAddress": "123 Partner Street",
                    "dateOfIncorporation": "2000-01-01",
                    "license": "ABC123",
                    "cin": "XYZ789",
                    "gstNo": "12AABCU1234K1ZV"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Partner signed up successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/otp/generate": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "generateOtp",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "partnerId": "12345"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP generated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/otp/verify": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "verifyOtp",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "partnerId": "1",
                    "emailOtp": "123456",
                    "mobileOtp": "654321"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP verification successful",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/activity": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "createActivity",
        "parameters": [
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "partnerId": "12345",
                    "activityEventType": "Sign-in",
                    "ipAddress": "192.168.1.1"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Partner activity created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/banks/activity/{partnerId}": {
      "post": {
        "tags": [
          "Partner-Controller"
        ],
        "operationId": "getUserActivityForSpecificPartner",
        "parameters": [
          {
            "name": "partnerId",
            "in": "path",
            "required": true,
            "description": "ID of the partner",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "in": "header",
            "description": "UUID for each request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RequestId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRequest"
              },
              "example": {
                "data": {
                  "type": "partner",
                  "attributes": {
                    "activityEventType": "Sign-in"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Activity data for a specific partner",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ResponseDTOObject": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DataDTOObject"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDTOObject"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
...
      "ResponseDTOListUserActivity": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DataDTOListUserActivity"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDTOListUserActivity"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
      "UserActivityRequest": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "attributes": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "pattern": "^[0-9]{1,6}$"
                  },
                  "eventType": {
                    "type": "string",
                    "pattern": "^[a-zA-Z]{5,50}$"
                  },
                  "eventDetails": {
                    "type": "string",
                    "pattern": "^[a-zA-Z]{5,100}$"
                  },
                  "startdate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "email": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9+_.-]+@(.+)$"
                  },
                  "mobile": {
                    "type": "string",
                    "pattern": "^[1-9]\\d{9}$"
                  }
                }
              }
            }
          }
        }
      },
      "RequestId": {
        "type": "string",
        "pattern": "^[a-zA-Z0-9]",
        "description": "UUID for each request"
      },
...
    },
...
  }
}